+++ app/core/typing.ts
@@ -1,5 +1,0 @@
-import { ChangesStreamService } from './service/ChangesStreamService';
-
-export interface ContextCnpmcore {
-  changesStreamService: ChangesStreamService;
-}