+++ app/port/typebox.ts
@@ -35,6 +35,10 @@ export const BinaryNameRule = Type.String({
   maxLength: 220,
 });
 
+// `[ -~]` matches all printable ASCII characters
+// https://catonmat.net/my-favorite-regex
+export const BinarySubpathRule = Type.RegEx(/^[ -~]{1,1024}$/);
+
 export const Tag = Type.String({
   format: 'semver-tag',
   transform: [ 'trim' ],